[FEAT] Added loader on the screen
